+++ archive.php
@@ -1,26 +1,47 @@
+<?php
+/**
+* The template for displaying Archive pages.
+ *
+ * Used to display archive-type pages if nothing more specific matches a query.
+ * For example, puts together date-based pages if no date.php file exists.
+ *
+ * If you'd like to further customize these archive views, you may create a
+ * new template file for each specific one. For example, Twenty Twelve already
+ * has tag.php for Tag archives, category.php for Category archives, and
+ * author.php for Author archives.
+ *
+ * Learn more: http://codex.wordpress.org/Template_Hierarchy
+
+ *
+ * @package WordPress
+ * @subpackage NARGA Framework
+ * @since Twenty Twelve 1.0
+ */
+?>
+
 <?php get_header(); ?>
-
-		<!-- Row for main content area -->
-		<div id="content" class="eight columns">
-	
-			<div class="post-box">
-				<h1>
-					<?php if (is_day()) : ?>
-						<?php printf(__('Daily Archives: %s', 'narga'), get_the_date()); ?>
-					<?php elseif (is_month()) : ?>
-						<?php printf(__('Monthly Archives: %s', 'narga'), get_the_date('F Y')); ?>
-					<?php elseif (is_year()) : ?>
-						<?php printf(__('Yearly Archives: %s', 'narga'), get_the_date('Y')); ?>
-					<?php else : ?>
-						<?php single_cat_title(); ?>
-					<?php endif; ?>
-				</h1>
-				<hr>
-				<?php get_template_part('loop', 'category'); ?>
-			</div>
-
-		</div><!-- End Content row -->
-		
-		<?php get_sidebar(); ?>
-		
+<!-- Row for main content area -->
+<div id="content" class="eight columns">
+<div class="post-box">
+<h2><?php if ( is_day() ) : printf( __( 'Daily Archives: %s', 'narga' ), '<span>' . get_the_date() . '</span>' );
+elseif ( is_month() ) :
+    printf( __( 'Monthly Archives: %s', 'narga' ), '<span>' . get_the_date( _x( 'F Y', 'monthly archives date format', 'narga' ) ) . '</span>' );
+elseif ( is_year() ) :
+    printf( __( 'Yearly Archives: %s', 'narga' ), '<span>' . get_the_date( _x( 'Y', 'yearly archives date format', 'narga' ) ) . '</span>' );
+else :
+    single_cat_title();
+endif;?></h2>
+<hr>
+<?php while (have_posts()) : the_post();
+get_template_part('content', get_post_format());
+endwhile; // End the loop ?>
+<?php /* Display navigation to next/previous pages when applicable */
+if ( function_exists('narga_pagination') ) { narga_pagination(); } else if ( is_paged() ) { ?>
+        <nav id="post-nav">
+        <div class="post-previous"><?php next_posts_link( __( '&larr; Older posts', 'narga' ) ); ?></div>
+        <div class="post-next"><?php previous_posts_link( __( 'Newer posts &rarr;', 'narga' ) ); ?></div>
+        </nav>
+        <?php } ?></div>
+</div><!-- End Content row -->
+<?php get_sidebar(); ?>
 <?php get_footer(); ?>