+++ comments.php
@@ -1,7 +1,26 @@
 <?php
+/**
+ * The template for displaying Comments.
+ *
+ * The area of the page that contains both current comments
+ * and the comment form. The actual display of comments is
+ * handled by a callback to twentytwelve_comment() which is
+ * located in the functions.php file.
+ *
+ * @package WordPress
+ * @subpackage NARGA Framework
+ * @since Twenty Twelve 1.0
+ */
+?>
+<?php
 // Do not delete these lines
 if (!empty($_SERVER['SCRIPT_FILENAME']) && 'comments.php' == basename($_SERVER['SCRIPT_FILENAME']))
     die (__('Please do not load this page directly. Thanks!', 'narga'));
+/*
+ * If the current post is protected by a password and
+ * the visitor has not yet entered the password we will
+ * return early without loading the comments.
+ */
 
 if ( post_password_required() ) { ?>
         <section id="comments">