+++ content-page.php
@@ -1,8 +1,17 @@
-<?php /* Start loop */ ?>
+<?php
+/**
+ * The template used for displaying page content in page.php
+ *
+ * @package WordPress
+ * @subpackage NARGA Framework
+ * @since NARGA Framework 1.2
+ */
+?>
+
 <?php while (have_posts()) : the_post(); ?>
 <article id="post-<?php the_ID(); ?>" <?php post_class(); ?>>
     <header>
-        <h2 class="entry-title"><?php the_title(); ?></h2>
+        <h1><a href="<?php the_permalink(); ?>" title="<?php echo esc_attr( sprintf( __( 'View %s', 'narga' ), the_title_attribute( 'echo=0' ) ) ); ?>" rel="bookmark"><?php the_title(); ?></a></h1>
     </header>
     <section class="entry-content">
     <?php the_content(); ?>