+++ footer.php
@@ -1,3 +1,16 @@
+<?php
+/**
+ * The template for displaying the footer.
+ *
+ * Contains footer content and the closing of the
+ * #main and #page div elements.
+ *
+ * @package WordPress
+ * @subpackage NARGA Framework
+ * @since NARGA Framework 1.0
+ */
+?>
+
 <!-- End row for main content area -->
                 </section>		
        </div>