+++ page.php
@@ -2,7 +2,7 @@
     <!-- Row for main content area -->
     <div id="content" class="eight columns page-content-wrapper">
     <div class="post-box">
-    <?php get_template_part('loop', 'page'); ?>
+    <?php get_template_part('content', 'page'); ?>
     </div>
     </div><!-- End Content row -->
 <?php get_sidebar(); ?>