+++ search.php
@@ -1,15 +1,19 @@
+<?php
+/**
+ * The template for displaying Search Results pages.
+ * @package WordPress
+ * @subpackage NARGA Framework
+ * @since NARGA Framework 1.0
+ */
+?>
+
 <?php get_header(); ?>
-
-		<!-- Row for main content area -->
-		<div id="content" class="eight columns">
-	
-			<div class="post-box">
-				<h1><?php _e('Search Results for', 'narga'); ?> "<?php echo get_search_query(); ?>"</h1>
-				<?php get_template_part('loop', 'search'); ?>
-			</div>
-
-		</div><!-- End Content row -->
-		
-		<?php get_sidebar(); ?>
-		
+<!-- Row for main content area -->
+    <div id="content" class="eight columns">
+    <div class="post-box">
+	<h2><?php _e('Search Results for', 'narga'); ?> "<?php echo get_search_query(); ?>"</h2>
+        <?php get_template_part('content', 'search'); ?>
+    </div>
+    </div><!-- End Content row -->
+<?php get_sidebar(); ?>
 <?php get_footer(); ?>