+++ sidebar.php
@@ -1,3 +1,15 @@
+<?php
+/**
+ * The sidebar containing the main widget area.
+ *
+ * If no active widgets in sidebar, let's hide it completely.
+ *
+ * @package WordPress
+ * @subpackage NARGA Framework
+ * @since NARGA Framework 1.0
+ */
+?>
+
 <aside id="sidebar" class="four columns" role="complementary">
 	<?php dynamic_sidebar("Sidebar"); ?>
 </aside><!-- /#sidebar -->