+++ templates/full-width.php
@@ -6,7 +6,7 @@ get_header(); ?>
     <!-- Row for main content area -->
     <div id="content" class="twelve columns full-width-content-wrapper">
     <div class="post-box">
-    <?php get_template_part('loop', 'page'); ?>
+    <?php get_template_part('content', 'page'); ?>
     </div>
     </div><!-- End Content row -->
 <?php get_footer(); ?>